feat: refactor plugin functionality into more files
