+++ src/constants.ts
@@ -1,2 +1,2 @@
 export const STORYBOOK_ANGULAR_ARG_TYPES = "STORYBOOK_ANGULAR_ARG_TYPES";
-export const STORYBOOK_COMPONENT_UUID = "__STORYBOOK_COMPONENT_HASH";
+export const STORYBOOK_COMPONENT_ID = "__STORYBOOK_COMPONENT_ID";
