+++ src/webpack-angular-types-plugin/templating/code-doc-dependency.ts
