+++ src/webpack-angular-types-plugin/templating/component-uuid-template.ts
@@ -1,0 +1,5 @@
+import { STORYBOOK_COMPONENT_ID } from "../../constants";
+
+export function getPrototypeUUIDCodeBlock(className: string, uuid: string) {
+    return `${className}.prototype["${STORYBOOK_COMPONENT_ID}"] = "${uuid}"`;
+}
